+++ src/PluginInfo.cs
@@ -4,5 +4,5 @@ internal static class PluginInfo
 {
     public const string Guid = "raisery.noblemod";
     public const string Name = "NobleMod";
-    public const string Version = "1.1.0";
+    public const string Version = "1.1.1";
 }